ch(optimize code)

+++ server/routes/products.js
@@ -1,10 +1,10 @@
 import express from 'express';
 import {
-  ensureToken, verifyProductInput,
+  ensureToken, verifyProductInput, isAdmin,
 } from '../helpers/validators';
 import sendResponse from '../helpers/responses';
 import {
-  getUser, createProduct, getProducts, deleteProducts, updateProducts,
+  createProduct, getProducts, deleteProducts, updateProducts,
 } from '../crud/db-query';
 import constants from '../helpers/constants';
 
@@ -13,25 +13,19 @@ const productsRouter = express.Router();
 
 
 productsRouter.post('/', verifyProductInput, ensureToken, (req, res) => {
-  if (req.body.decoded.level !== constants.adminLevel) {
+  if (!isAdmin(req.body.decoded.level)) {
     sendResponse(res, 403, null, 'Your are not authorized to modify this content');
     return;
   }
   // query database
-  getUser(req.body.decoded.email)
-    .then(() => {
-      createProduct(req.body).then(() => {
-        sendResponse(res, 201, `'${req.body.productName}' was successfully added`);
-      }).catch((e) => {
-        console.log(e);
-        sendResponse(res, 500, null, 'Internal server error');
-      });
-      // res.send(result);
-    })
-    .catch((e) => {
-      console.log(e);
-      sendResponse(res, 404, null, 'Invalid user');
-    });
+
+  createProduct(req.body).then(() => {
+    sendResponse(res, 201, `'${req.body.productName}' was successfully added`);
+  }).catch(() => {
+    // console.log(e);
+    sendResponse(res, 500, null, 'Internal server error');
+  });
+  // res.send(result);
 });
 
 productsRouter.get('/', ensureToken, (req, res) => {
@@ -42,8 +36,8 @@ productsRouter.get('/', ensureToken, (req, res) => {
       message: 'successfully fetched products',
       productsArray: result,
     });
-  }).catch((e) => {
-    console.log(e);
+  }).catch(() => {
+    // console.log(e);
     sendResponse(res, 500, null, 'Internal server error');
   });
 });
@@ -60,22 +54,26 @@ productsRouter.get('/:id', ensureToken, (req, res) => {
     } else {
       sendResponse(res, 404, null, 'Product not found');
     }
-  }).catch((e) => {
-    console.log(e);
+  }).catch(() => {
+    // console.log(e);
     sendResponse(res, 500, null, 'Internal server error');
   });
 });
 
 productsRouter.delete('/:id', ensureToken, (req, res) => {
   // query database
+  if (!isAdmin(req.body.decoded.level)) {
+    sendResponse(res, 403, null, 'Your are not authorized to modify this content');
+    return;
+  }
   deleteProducts(req.params.id).then((result) => {
     if (result) {
       sendResponse(res, 200, 'successfully deleted product');
     } else {
       sendResponse(res, 404, null, 'Product not found');
     }
-  }).catch((e) => {
-    console.log(e);
+  }).catch(() => {
+    // console.log(e);
     sendResponse(res, 500, null, 'Internal server error');
   });
 });
@@ -94,8 +92,8 @@ productsRouter.put('/:id', verifyProductInput, ensureToken, (req, res) => {
     } else {
       sendResponse(res, 404, null, 'Invalid product');
     }
-  }).catch((e) => {
-    console.log(e);
+  }).catch(() => {
+    // console.log(e);
     sendResponse(res, 500, null, 'Internal server error');
   });
 });
